notify_anomalies/requirements.txt
pylint
pytest
python-dotenv
pyarrow
pandas
pymssql
boto3